refactor(esp32): rename ble_device_info to ble_dis for cleaner naming
- Rename ble_device_info.c/.h to ble_dis.c/.h
- Update all includes from "ble_device_info.h" to "ble_dis.h"
- Update CMakeLists.txt to reference new filenames
- No functional changes, purely naming cleanup

Rationale: "ble_dis" is more concise and aligns with standard BLE terminology:
- DIS = Device Information Service (standard BLE service)
- Matches naming pattern with other BLE services (ble_nus, ble_bas)
- Shorter, cleaner naming convention

ESP32 BLE architecture now uses consistent naming:
├── ble_nus.c/h    - Nordic UART Service client
├── ble_dis.c/h    - Device Information Service client
└── ble_bas.c/h    - Battery Service client
