+++ esp32/main/CMakeLists.txt
@@ -1,3 +1,3 @@
-idf_component_register(SRCS "ble_nus.c" "ble_device_info.c" "button.c" "nus_cdc_bridge.c" "bootloader_trigger.c" "ble_bas.c" "leds.c" "main.c" "ble_transport.c" "usb_hid.c" "usb_cdc.c"
+idf_component_register(SRCS "ble_dis.c" "ble_nus.c" "button.c" "nus_cdc_bridge.c" "bootloader_trigger.c" "ble_bas.c" "leds.c" "main.c" "ble_transport.c" "usb_hid.c" "usb_cdc.c"
                        INCLUDE_DIRS "."
                        REQUIRES bt esp_hid nvs_flash esp_driver_uart)
